# Allow Modifying Transaction in Dry Run and Other Improvements