--- a/Makefile
+++ b/Makefile
@@ -26,8 +26,9 @@ blackbox-smoke-prefix:
 	ls -l sandbox
 	cd sandbox && docker-compose ps
 
+NUM_PROCS = auto
 integration-test:
-	pytest -sv tests/integration
+	pytest -n $(NUM_PROCS) --durations=10 -sv tests/integration
 
 pre-commit-check: lint unit-test integration-test