--- a/Makefile
+++ b/Makefile
@@ -37,7 +37,7 @@ integration-test:
 	pytest -n $(NUM_PROCS) --durations=10 -sv tests/integration
 
 notebooks-test:
-	pytest --nbmake notebooks
+	pytest --nbmake -n $(NUM_PROCS) notebooks
 
 pre-commit-check: lint unit-test integration-test