--- a/Makefile
+++ b/Makefile
@@ -57,6 +57,11 @@ local-blackbox-smoke: blackbox-smoke-prefix local-sandbox-test
 
 local-blackbox: local-blackbox-smoke integration-test
 
+NOTEBOOK = notebooks/quadratic_factoring_game.ipynb
+# assumes already ran `make pip-notebooks`
+local-notebook: 
+	 jupyter retro $(NOTEBOOK)
+
 # assumes act is installed, e.g. via `brew install act`:
 local-gh-simulate:
 	act