--- a/graviton/dryrun.py
+++ b/graviton/dryrun.py
@@ -549,6 +549,7 @@ class DryRunHelper:
 
     @classmethod
     def _app_request(cls, program, app, txn_params):
+        """TODO: out to be able to stop delegating at tis point"""
         run_mode = cls._get_run_mode(app)
         enriched = cls._prepare_app(app)
         txn = transaction.ApplicationCallTxn(**txn_params)
@@ -556,6 +557,7 @@ class DryRunHelper:
 
     @classmethod
     def _lsig_request(cls, program, lsig, txn_params):
+        """TODO: out to be able to stop delegating at tis point"""
         enriched = cls._prepare_lsig(lsig)
         txn = transaction.PaymentTxn(**txn_params)
         return cls._prepare_lsig_source_request(program, enriched, txn)