## Summary of Changes
* `.github/workflows/build.yml` + `Makefile`:
  * allow testing and running jupyter notebooks
* `graviton/blackbox.py`:
  * pass `mypy`
  * allow specifying transaction params into dry-runs
* `graviton/dryrun.py`:
  * pass `mypy`
* `graviton/deprecated_dry_run.py`: new module that contains functionality mostly used by the now deprecated `DryRunTestCaseMixin`
* `graviton/deprecated_dryrun_mixin.py`: where the now deprecated `DryRunTestCaseMixin` is defined
* `graviton/dryrun.py`:
  * refactor to keep only the most useful portions of the original helper
  * make `mypy` pass
  * allow specifying transaction parameters
  * `pprint()` now returns a string, in addition to printing
* `graviton/invariant.py`:  make `mypy` pass
* `graviton/models.py`: new module that contains lightweight models such as the `App` and `LSig` dataclasses
* `mypy.ini`: needed by mypy
* `notebooks`: new directory for Jupyter notebooks. It contains:
  * `IMAGES.md`: a static file that shows examples of 3D graphs of dry-runs
  * `notebooks/notebooks_test.ipynb`: explains how to test notebooks with `pytest`
  * `notebooks/quadratic_factoring_game.ipynb`: Blackbox demo
* `setup.py`: new Jupyter notebooks requirements, and declaration for receiving projects that this package is typed
* `tests/integration/algod_test.py`: separate out the test that `algod` is alive and well into a seperate integration test, for easier debugging
* `tests/integration/algod_test.py`: small refactoring, mostly for `mypy`
* `tests/integration/dryrun_mixin_docs_test.py`: refactoring mostly due to deprecations
* `tests/integration/lsig_test.py`: New logic sig only test. THIS NEEDS A CLEANUP.

## adding `pytest-xdist` for faster integration testing.

### Comparison on github:

* [Before](https://github.com/algorand/graviton/runs/5880231112?check_suite_focus=true#step:11:3486)
 (58.19 secs for python 3.10)
* [After](https://github.com/algorand/graviton/runs/5886783616?check_suite_focus=true#step:11:2696)
 (37.64 secs for python 3.10)

### Local test comparison:

Runniing with a single thread:
```sh
❯ make integration-test NUM_PROCS=1

 ... 

================================================================== slowest 10 durations ===================================================================
16.47s call     tests/integration/blackbox_test.py::test_app_with_report[app_slow_fibonacci]
8.85s call     tests/integration/blackbox_test.py::test_logicsig_with_report[lsig_slow_fibonacci]
5.03s call     tests/integration/blackbox_test.py::test_logicsig_with_report[lsig_string_mult]
4.83s call     tests/integration/blackbox_test.py::test_app_with_report[app_string_mult]
4.70s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V4]
4.25s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V3]
3.90s call     tests/integration/lsig_test.py::test_factorizer_report_with_pymnt
3.56s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V2]
2.43s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V1]
1.49s call     tests/integration/dryrun_mixin_docs_test.py::ExampleTestCase::test_factorial
========= 1737 passed in 82.38s (0:01:22) ========= 
```

With 
```sh
❯ make integration-test

...

================================================================== slowest 10 durations ===================================================================
38.58s call     tests/integration/blackbox_test.py::test_app_with_report[app_slow_fibonacci]
25.53s call     tests/integration/blackbox_test.py::test_logicsig_with_report[lsig_slow_fibonacci]
19.45s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V4]
18.16s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V3]
18.12s call     tests/integration/blackbox_test.py::test_app_with_report[app_string_mult]
16.77s call     tests/integration/blackbox_test.py::test_logicsig_with_report[lsig_string_mult]
6.77s call     tests/integration/dryrun_mixin_docs_test.py::ExampleTestCase::test_factorial
5.49s call     tests/integration/lsig_test.py::test_factorizer_report_with_pymnt
4.59s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V1]
3.27s call     tests/integration/lsig_test.py::test_factorizer_game_version_report[V2]
========= 1737 passed in 44.31s =========
```